requirements.txt
Flask==2.2.3
py-cpuinfo==9.0.0
psutil==5.9.4
numpy==1.24.2
WMI==1.5.1
